requirements.txt
fastapi
uvicorn
python-multipart
pillow
aiofiles
jinja2